renamed listItems.js to sideBarItems.js and updated import
